pyproject.toml
[project]
name = "word-segmenter"
version = "0.1.0"
description = "Automate Azerbaijani morphology segmentation and analysis using Google Gemini LLM and Python."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-generativeai>=0.8.5",
    "pydantic>=2.11.4",
]
